hackaton-cepedi-back-ia/requirements.txt
asgiref==3.8.1
black==24.3.0
click==8.1.7
Django==5.0.3
djangorestframework==3.15.1
joblib==1.3.2
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
python-decouple==3.8
scikit-learn==1.3.0
scipy==1.12.0
sqlparse==0.4.4
threadpoolctl==3.4.0
tomli==2.0.1
typing_extensions==4.10.0
